+++ research/src/main/java/com/research/garage/repository/ZipCodeRepository.java
@@ -80,15 +80,50 @@ public class ZipCodeRepository implements IZipCodeRepository {
 
 	@Override
 	public String deleteZipCodeById(String ZipCodeId) {
-		ZipCodeDAO.deleteById(ZipCodeId);
-		return "ZipCode Deleted";
+		if(ZipCodeDAO.getOne(ZipCodeId)!=null)
+		{
+			ZipCodeDAO.deleteById(ZipCodeId);
+			return "ZipCode Deleted";
+		}
+		else
+		{
+			return "ZipCode id Not found";
+		}
+		
 	}
 
 	@Override
-	public String UpdateZipCode(ZipCodeDTO ZipCodeDTO) {
-		ZipCodeProjection zipCode = ZipCodeDAO.getOne(ZipCodeDTO.getZipCodeId());
+	public String UpdateZipCode(ZipCodeDTO zipCodeDTO) {
+		ZipCodeProjection zipCode = ZipCodeDAO.getOne(zipCodeDTO.getZipCodeId());
 
-		zipCode.setCity_Name(ZipCodeDTO.getCityName());
+		CountryProjection country = countryDAO.getOne(zipCodeDTO.getCountryId());
+		StateProjection state = stateDAO.getOne(zipCodeDTO.getStateId());
+	//	ZipCodeProjection zipCode = new ZipCodeProjection();
+		if(zipCodeDTO.getZipCodeId()==null||zipCodeDTO.getZipCodeId().isEmpty())
+		{
+			return "Zip Code  "+ zipCodeDTO.getZipCodeId() +" Should Not Be Empty";
+		}
+		ZipCodeProjection zipCode1 = ZipCodeDAO.getOne(zipCodeDTO.getZipCodeId());
+		if(zipCode1!=null)
+		{
+			return "Zip Code  "+ zipCodeDTO.getZipCodeId() +" Already Available";
+		}
+		if(country==null)
+		{
+			return "Country ID "+ zipCodeDTO.getCountryId() +" not in the Master";
+		}
+		if(state==null)
+		{
+			return "State ID "+ zipCodeDTO.getStateId() +" not in the Master";
+		}
+		
+		
+		zipCode.setZip_Code_Id(zipCodeDTO.getZipCodeId());
+		zipCode.setCity_Name(zipCodeDTO.getCityName());
+		zipCode.setCountry_Id(zipCodeDTO.getCountryId());
+		zipCode.setState_Id(zipCodeDTO.getStateId());
+
+		ZipCodeDAO.save(zipCode);
 
 		return "ZipCode Updated";
 	}
